Add tests

+++ programs/fusion-swap/src/auction_utils.rs
@@ -15,29 +15,29 @@ pub struct DutchAuctionData {
     pub points_and_time_deltas: Vec<PointsAndTimeDeltas>,
 }
 
-pub fn calculate_rate_bump(timestamp: u32, data: DutchAuctionData) -> u32 {
+pub fn calculate_rate_bump(timestamp: u32, data: DutchAuctionData) -> u64 {
     if timestamp <= data.auction_start_time {
-        return data.initial_rate_bump;
+        return data.initial_rate_bump as u64;
     }
     if timestamp >= data.auction_finish_time {
         return 0;
     }
 
-    let mut current_rate_bump = data.initial_rate_bump;
-    let mut current_point_time = data.auction_start_time;
+    let mut current_rate_bump = data.initial_rate_bump as u64;
+    let mut current_point_time = data.auction_start_time as u64;
 
     for point_and_time_delta in data.points_and_time_deltas.iter() {
-        let next_rate_bump = point_and_time_delta.rate_bump;
-        let next_point_time = current_point_time + point_and_time_delta.point_time as u32;
-        if timestamp <= next_point_time {
-            return ((timestamp - current_point_time) * next_rate_bump
-                + (next_point_time - timestamp) * current_rate_bump)
+        let next_rate_bump = point_and_time_delta.rate_bump as u64;
+        let next_point_time = current_point_time + point_and_time_delta.point_time as u64;
+        if timestamp as u64 <= next_point_time {
+            return ((timestamp as u64 - current_point_time) * next_rate_bump
+                + (next_point_time - timestamp as u64) * current_rate_bump)
                 / (next_point_time - current_point_time);
         }
 
         current_rate_bump = next_rate_bump;
         current_point_time = next_point_time;
     }
-    (data.auction_finish_time - timestamp) * current_rate_bump
-        / (data.auction_finish_time - current_point_time)
+    (data.auction_finish_time as u64 - timestamp as u64) * current_rate_bump
+        / (data.auction_finish_time as u64 - current_point_time)
 }
